+++ headers/cg/io/rectangle.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "cg/primitives/rectangle.h"
+#include "range.h"
+
+namespace cg
+{
+   std::istream & operator >> (std::istream & in, rectangle_2t<Scalar> & rect)
+   {
+      using io::skip_char;
+
+      return skip_char(skip_char(skip_char(in, '(') >> rect.x, ',') >> rect.y, ')');
+   }
+
+   std::ostream & operator << (std::ostream & out, rectangle_2t<Scalar> const & rect)
+   {
+      out << "(" << rect.x << ", " << rect.y << ")";
+      return out;
+   }
+}
