--- headers/io/range.h
@@ -1,13 +1,0 @@
-#ifndef _RANGE_IO_H_
-#define _RANGE_IO_H_
-
-#include "geom/primitives/range.h"
-
-namespace geom {
-namespace structures {
-
-    std::ostream & operator << (std::ostream & out, range_f const &);
-
-}}
-
-#endif /* _RANGE_IO_H_ */
