+++ src/visualization/drawer_impl.h
@@ -1,5 +1,4 @@
-#ifndef _DRAWER_IMPL_H_
-#define _DRAWER_IMPL_H_
+#pragma once
 
 #include <QtOpenGL>
 
@@ -7,41 +6,37 @@
 
 namespace visualization
 {
+   struct drawer_impl : drawer_type
+   {
+      void set_color(QColor const & c);
+      void draw_line(segment_2f const &, float width);
+      void draw_line(point_2f const &, point_2f const &, float width);
+      void draw_point(point_2f const & pt, float radius);
 
-struct drawer_impl : drawer_type
-{
-    void set_color(QColor const & c);
-    void draw_line(segment_2f const &, float width);
-    void draw_line(point_2f const &, point_2f const &, float width);
-    void draw_point(point_2f const & pt, float radius);
+      drawer_impl()
+         : current_color_ (Qt::black)
+      {}
 
-    drawer_impl()
-        : current_color_ (Qt::black)
-    {}
+      void clear();
 
-    void clear();
+      struct point_buffer_t
+      {
+         std::vector<GLfloat>    points;
+         std::vector<GLfloat>    colors;
+         float radius;
+      };
 
-    struct point_buffer_t
-    {
-        std::vector<GLfloat>    points;
-        std::vector<GLfloat>    colors;
-        float radius;
-    };
+      struct segment_buffer_t
+      {
+         std::vector<GLfloat>    segments;
+         std::vector<GLfloat>    colors;
+         float width;
+      };
 
-    struct segment_buffer_t
-    {
-        std::vector<GLfloat>    segments;
-        std::vector<GLfloat>    colors;
-        float width;
-    };
+      std::vector<point_buffer_t>     point_buffers;
+      std::vector<segment_buffer_t>   segment_buffers;
 
-    std::vector<point_buffer_t>     point_buffers;
-    std::vector<segment_buffer_t>   segment_buffers;
-
-private:
-    QColor current_color_;
-};
-
+   private:
+      QColor current_color_;
+   };
 }
-
-#endif /*_DRAWER_IMPL_H_*/
